Move BufferLine into core
