+++ src/SelectionManager.test.ts
@@ -11,7 +11,7 @@ import { BufferSet } from './BufferSet';
 import { ITerminal, IBuffer } from './Types';
 import { IBufferLine } from './core/Types';
 import { MockTerminal } from './TestUtils.test';
-import { BufferLine, CellData } from './BufferLine';
+import { BufferLine, CellData } from './core/buffer/BufferLine';
 
 class TestMockTerminal extends MockTerminal {
   emit(event: string, data: any): void {}
