+++ src/core/buffer/BufferLine.ts
@@ -2,9 +2,9 @@
  * Copyright (c) 2018 The xterm.js authors. All rights reserved.
  * @license MIT
  */
-import { CharData, IBufferLine, ICellData, IColorRGB, IAttributeData } from './core/Types';
-import { stringFromCodePoint } from './core/input/TextDecoder';
-import { DEFAULT_COLOR } from './common/Types';
+import { CharData, IBufferLine, ICellData, IColorRGB, IAttributeData } from '../Types';
+import { stringFromCodePoint } from '../input/TextDecoder';
+import { DEFAULT_COLOR } from '../../common/Types';
 
 export const DEFAULT_ATTR = (0 << 18) | (DEFAULT_COLOR << 9) | (256 << 0);
 
@@ -325,17 +325,15 @@ export class CellData extends AttributeData implements ICellData {
  * memory allocs / GC pressure can be greatly reduced by reusing the CellData object.
  */
 export class BufferLine implements IBufferLine {
-  protected _data: Uint32Array | null = null;
+  protected _data: Uint32Array;
   protected _combined: {[index: number]: string} = {};
   public length: number;
 
   constructor(cols: number, fillCellData?: ICellData, public isWrapped: boolean = false) {
-    if (cols) {
-      this._data = new Uint32Array(cols * CELL_SIZE);
-      const cell = fillCellData || CellData.fromCharData([0, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE]);
-      for (let i = 0; i < cols; ++i) {
-        this.setCell(i, cell);
-      }
+    this._data = new Uint32Array(cols * CELL_SIZE);
+    const cell = fillCellData || CellData.fromCharData([0, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE]);
+    for (let i = 0; i < cols; ++i) {
+      this.setCell(i, cell);
     }
     this.length = cols;
   }
@@ -573,7 +571,7 @@ export class BufferLine implements IBufferLine {
           }
         }
       } else {
-        this._data = null;
+        this._data = new Uint32Array(0);
         this._combined = {};
       }
     }
