+++ src/renderer/CharacterJoinerRegistry.test.ts
@@ -10,7 +10,7 @@ import { CircularList } from '../common/CircularList';
 
 import { ICharacterJoinerRegistry } from './Types';
 import { CharacterJoinerRegistry } from './CharacterJoinerRegistry';
-import { BufferLine, CellData } from '../BufferLine';
+import { BufferLine, CellData } from '../core/buffer/BufferLine';
 import { IBufferLine } from '../core/Types';
 
 describe('CharacterJoinerRegistry', () => {
