+++ src/renderer/CursorRenderLayer.ts
@@ -7,7 +7,7 @@ import { IColorSet, IRenderDimensions } from './Types';
 import { BaseRenderLayer } from './BaseRenderLayer';
 import { ITerminal } from '../Types';
 import { ICellData } from '../core/Types';
-import { CellData } from '../BufferLine';
+import { CellData } from '../core/buffer/BufferLine';
 
 interface ICursorState {
   x: number;
